+++ packages/react/src/components/List/ListItem/ListItem.jsx
@@ -66,6 +66,8 @@ const ListItemPropTypes = {
   isDragging: PropTypes.bool.isRequired,
   onItemMoved: PropTypes.func.isRequired,
   itemWillMove: PropTypes.func.isRequired,
+  /** true if the list item should not be focusable even though isSelectable is true */
+  preventRowFocus: PropTypes.bool,
 };
 
 const ListItemDefaultProps = {
@@ -94,6 +96,7 @@ const ListItemDefaultProps = {
   },
   selectedItemRef: null,
   tags: null,
+  preventRowFocus: false,
 };
 
 const ListItem = ({
@@ -125,6 +128,7 @@ const ListItem = ({
   connectDragPreview,
   itemWillMove,
   dragPreviewText,
+  preventRowFocus,
 }) => {
   const mergedI18n = useMemo(() => ({ ...ListItemDefaultProps.i18n, ...i18n }), [i18n]);
 
@@ -250,6 +254,7 @@ const ListItem = ({
         itemWillMove,
         disabled,
         renderDropTargets,
+        preventRowFocus,
       }}
     >
       {renderDragPreview()}