+++ packages/react/src/components/SideNav/SideNav.test.jsx
@@ -1,12 +1,13 @@
 import React from 'react';
-import { Switcher24, Chip24, Group24 } from '@carbon/icons-react';
+import { Switcher24, Chip24, Group24, ParentChild24 } from '@carbon/icons-react';
 import { render, screen } from '@testing-library/react';
+import userEvent from '@testing-library/user-event';
 
 import { settings } from '../../constants/Settings';
 
 import SideNav from './SideNav';
 
-const { prefix } = settings;
+const { prefix, iotPrefix } = settings;
 
 React.Fragment = ({ children }) => children;
 
@@ -216,5 +217,151 @@ describe('SideNav', () => {
     expect(screen.getByText('Members').closest('li')).toHaveClass(
       `${prefix}--side-nav__item--active`
     );
+  });
+
+  it('should render nested levels of children', () => {
+    const onClick = jest.fn();
+    render(
+      <SideNav
+        {...mockProps}
+        links={[
+          {
+            isEnabled: true,
+            icon: ParentChild24,
+            metaData: {
+              label: 'Nested Levels',
+              element: 'button',
+            },
+            linkContent: 'Nested Levels',
+            childContent: [
+              {
+                metaData: {
+                  label: 'Co-Parent Link',
+                  title: 'Co-Parent Link',
+                  element: 'a',
+                  href: 'https://www.ibm.com',
+                },
+                content: 'Co-Parent Link',
+              },
+              {
+                metaData: {
+                  label: 'Parent',
+                  title: 'Parent',
+                  element: 'button',
+                },
+                content: 'Parent',
+                linkContent: 'Parent',
+                childContent: [
+                  {
+                    metaData: {
+                      label: 'Sibling 1 Link',
+                      title: 'Sibling 1 Link',
+                      element: 'a',
+                      href: 'https://www.ibm.com',
+                    },
+                    content: 'Sibling 1 Link',
+                  },
+                  {
+                    isEnabled: true,
+                    metaData: {
+                      label: 'Child',
+                      element: 'button',
+                    },
+                    linkContent: 'Child',
+                    childContent: [
+                      {
+                        metaData: {
+                          label: 'Grandchild Button',
+                          title: 'Grandchild Button',
+                          onClick,
+                          element: 'button',
+                        },
+                        content: 'Grandchild Button',
+                        isActive: true,
+                      },
+                      {
+                        metaData: {
+                          label: 'Grandchild Link',
+                          title: 'Grandchild Link',
+                          href: 'https://www.ibm.com',
+                          element: 'a',
+                        },
+                        content: 'Grandchild Link',
+                      },
+                    ],
+                  },
+                  {
+                    metaData: {
+                      label: 'Sibling 2 Button',
+                      title: 'Sibling 2 Button',
+                      element: 'button',
+                      onClick,
+                    },
+                    content: 'Sibling 2 Button',
+                  },
+                ],
+              },
+              {
+                metaData: {
+                  label: 'Co-Parent Button',
+                  title: 'Co-Parent Button',
+                  element: 'button',
+                  onClick,
+                },
+                content: 'Co-Parent Button',
+              },
+            ],
+          },
+        ]}
+      />
+    );
+
+    const expectToBeVisible = (text) => {
+      expect(screen.getByText(text)).toBeVisible();
+    };
+
+    const clickText = (text) => {
+      userEvent.click(screen.getByText(text));
+    };
+
+    const expectToBeActive = (text) => {
+      expect(screen.getByText(text).closest(`.${prefix}--side-nav__item`)).toHaveClass(
+        `${prefix}--side-nav__item--active`
+      );
+    };
+
+    const expectToHaveDepth = (text, depth) => {
+      expect(screen.getByText(text).closest(`.${prefix}--side-nav__item`)).toHaveClass(
+        `${iotPrefix}--side-nav__item--depth-${depth}`
+      );
+    };
+
+    expectToBeVisible('Nested Levels');
+    expectToBeActive('Nested Levels');
+    expectToHaveDepth('Nested Levels', 0);
+
+    clickText('Nested Levels');
+    expectToBeVisible('Co-Parent Link');
+    expectToBeVisible('Parent');
+    expectToBeVisible('Co-Parent Button');
+    clickText('Co-Parent Button');
+    expect(onClick).toHaveBeenCalledTimes(1);
+    expectToBeActive('Parent');
+    expectToHaveDepth('Parent', 1);
+
+    clickText('Parent');
+    expectToBeVisible('Sibling 1 Link');
+    expectToBeVisible('Child');
+    expectToBeVisible('Sibling 2 Button');
+    clickText('Sibling 2 Button');
+    expect(onClick).toHaveBeenCalledTimes(2);
+    expectToBeActive('Child');
+    expectToHaveDepth('Child', 2);
+
+    clickText('Child');
+    expectToBeVisible('Grandchild Button');
+    expectToBeVisible('Grandchild Link');
+    clickText('Grandchild Button');
+    expect(onClick).toHaveBeenCalledTimes(3);
   });
 });